+++ src/hooks/useMsgInput.ts
@@ -1,9 +1,6 @@
-import { LimitEnum, MittEnum, MsgEnum, MessageStatusEnum } from '@/enums'
+import { LimitEnum, MittEnum, MsgEnum } from '@/enums'
 import { useUserInfo } from '@/hooks/useCached.ts'
-import apis from '@/services/apis.ts'
 import { useCachedStore, type BaseUserItem } from '@/stores/cached.ts'
-import { useChatStore } from '@/stores/chat.ts'
-import { useGlobalStore } from '@/stores/global.ts'
 import { useSettingStore } from '@/stores/setting.ts'
 import { useMitt } from '@/hooks/useMitt.ts'
 import { type } from '@tauri-apps/plugin-os'
@@ -12,13 +9,11 @@ import { Ref } from 'vue'
 import { useCommon } from './useCommon.ts'
 import { readText, readImage } from '@tauri-apps/plugin-clipboard-manager'
 import Database from '@tauri-apps/plugin-sql'
-import { messageStrategyMap } from '@/strategy/MessageStrategy.ts'
+import { messageStrategyMap, resetInput } from '@/strategy/MessageStrategy.ts'
 import { useTrigger } from './useTrigger'
 import type { AIModel } from '@/services/types.ts'
 
 export const useMsgInput = (messageInputDom: Ref) => {
-  const chatStore = useChatStore()
-  const globalStore = useGlobalStore()
   const cachedStore = useCachedStore()
   const { triggerInputEvent, insertNode, getMessageContentType, getEditorRange, imgPaste, reply, userUid } = useCommon()
   const settingStore = useSettingStore()
@@ -188,41 +183,6 @@ export const useMsgInput = (messageInputDom: Ref) => {
     chat.value.sendKey = v
   })
 
-  /**
-   * 从HTML内容中提取 @ 用户的uid
-   * @param content HTML格式的消息内容
-   * @param userList 用户列表
-   * @returns 被 @ 用户的uid数组
-   */
-  const extractAtUserIds = (content: string, userList: BaseUserItem[]): number[] => {
-    const atUserIds: number[] = []
-
-    // 创建临时DOM元素来解析HTML
-    const tempDiv = document.createElement('div')
-    tempDiv.innerHTML = content
-
-    // 获取纯文本内容
-    const textContent = tempDiv.textContent || ''
-
-    // 使用更精确的正则表达式匹配@用户
-    // 匹配@后面的非空白字符，直到遇到空白字符或字符串结束
-    const regex = /@([^\s]+)/g
-    const matches = textContent.match(regex)
-
-    if (matches) {
-      matches.forEach((match) => {
-        const username = match.slice(1) // 移除@符号
-        const user = userList.find((u) => u.name === username)
-        if (user) {
-          atUserIds.push(user.uid)
-        }
-      })
-    }
-
-    // 去重并返回
-    return [...new Set(atUserIds)]
-  }
-
   /** 去除html标签(用于鉴别回复时是否有输入内容) */
   const stripHtml = (html: string) => {
     try {
@@ -236,17 +196,6 @@ export const useMsgInput = (messageInputDom: Ref) => {
     } catch (error) {
       console.error('Error in stripHtml:', error)
       return ''
-    }
-  }
-
-  /** 重置输入框内容 */
-  const resetInput = () => {
-    try {
-      msgInput.value = ''
-      messageInputDom.value.innerHTML = ''
-      reply.value = { avatar: '', imgCount: 0, accountName: '', content: '', key: 0 }
-    } catch (error) {
-      console.error('Error in resetInput:', error)
     }
   }
 
@@ -272,81 +221,7 @@ export const useMsgInput = (messageInputDom: Ref) => {
       window.$message.warning('暂不支持发送类型消息')
       return
     }
-    const msg = messageStrategy.getMsg(msgInput.value, reply.value)
-
-    // 从消息内容中提取@用户的uid
-    const atUidList = extractAtUserIds(msgInput.value, cachedStore.currentAtUsersList)
-
-    // 创建临时消息ID
-    const tempMsgId = Date.now()
-    // 根据消息类型创建消息体
-    const messageBody = {
-      ...messageStrategy.buildMessageBody(msg, reply),
-      atUidList // 添加@用户列表
-    }
-
-    // 创建消息对象;
-    const tempMsg = messageStrategy.buildMessageType(tempMsgId, messageBody, globalStore, userUid)
-    // 先添加到消息列表
-    chatStore.pushMsg(tempMsg)
-
-    // 设置800ms后显示发送状态的定时器
-    const statusTimer = setTimeout(() => {
-      chatStore.updateMsg({
-        msgId: tempMsgId,
-        status: MessageStatusEnum.SENDING
-      })
-    }, 800)
-
-    console.log('发送消息', messageBody, msg.type)
-    apis
-      .sendMsg({
-        roomId: globalStore.currentSession.roomId,
-        msgType: msg.type,
-        body: messageBody
-      })
-      .then(async (res) => {
-        clearTimeout(statusTimer)
-        // 更新消息状态为成功，同时更新消息ID和回复内容
-        chatStore.updateMsg({
-          msgId: tempMsgId,
-          status: MessageStatusEnum.SUCCESS,
-          newMsgId: res.message.id,
-          body: res.message.body // 更新消息体，包含服务器返回的回复内容
-        })
-        if (res.message.type === MsgEnum.TEXT) {
-          await chatStore.pushMsg(res)
-          // 保存到数据库
-          await db.value?.execute(
-            'INSERT INTO message (room_id, from_uid, content, reply_msg_id, status, gap_count, type, create_time, update_time) VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9)',
-            [
-              globalStore.currentSession.roomId,
-              userUid.value,
-              msg.content,
-              msg.reply,
-              0,
-              0,
-              msg.type,
-              new Date().getTime(),
-              new Date().getTime()
-            ]
-          )
-        }
-        // 发完消息就要刷新会话列表，
-        //  FIXME 如果当前会话已经置顶了，可以不用刷新
-        chatStore.updateSessionLastActiveTime(globalStore.currentSession.roomId)
-      })
-      .catch(() => {
-        clearTimeout(statusTimer)
-        // 更新消息状态为失败
-        chatStore.updateMsg({
-          msgId: tempMsgId,
-          status: MessageStatusEnum.FAILED
-        })
-      })
-
-    // 清空输入框和回复信息
-    resetInput()
+    messageStrategy.send(msgInput, messageInputDom)
   }
 
   /** 当输入框手动输入值的时候触发input事件(使用vueUse的防抖) */
@@ -388,7 +263,7 @@ export const useMsgInput = (messageInputDom: Ref) => {
     if (disabledSend.value) {
       e.preventDefault()
       e.stopPropagation()
-      resetInput()
+      resetInput(msgInput, messageInputDom)
       return
     }
